requirements.txt
alembic==1.0.7
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.7.0
cryptography==2.5
cycler==0.10.0
dask==1.1.0
decorator==4.3.2
emoji==0.5.1
future==0.17.1
idna==2.8
kiwisolver==1.0.1
Mako==1.0.7
MarkupSafe==1.1.0
mysqlclient==1.4.1
networkx==2.2
Pillow==5.4.1
pycparser==2.19
pyparsing==2.3.1
python-dateutil==2.7.5
python-editor==1.0.3
python-Levenshtein==0.12.0
python-telegram-bot==12.0.0b1
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
scipy==1.2.0
six==1.12.0
SQLAlchemy==1.2.17
toolz==0.9.0
urllib3==1.24.1